clang format

+++ cpp/src/io/avro/avro.cpp
@@ -44,10 +44,11 @@ int64_t container::get_encoded()
 template <>
 std::string container::get_encoded()
 {
-  auto const len = [&] { 
+  auto const len = [&] {
     auto const len = get_encoded<uint64_t>();
-    return (len & 1) || (m_cur >= m_end) ? 0 : std::min(len >> 1, static_cast<size_t>(m_end - m_cur));
-  } ();
+    return (len & 1) || (m_cur >= m_end) ? 0
+                                         : std::min(len >> 1, static_cast<uint64_t>(m_end - m_cur));
+  }();
   auto const s = reinterpret_cast<char const *>(m_cur);
   m_cur += len;
   return std::string(s, len);
